Make error_getMsg take an error code, not an index
The error_getMsg function assumed that the user was passing not an error
code, but an error *index* into the internal ErrorMsgs struct.  This is
not how the error_getMsg function is called elsewhere.  For example, see the
`swmm_getAPIError` function.

Now when the user is handed an error code (a member of the ErrorCodes
struct) they can directly convert it to a useful message using
error_getMsg rather than having to know about where it lives inside the
ErrorMsgs struct.